# deploy/postman/noema-openapi.yaml
openapi: "3.0.3"
info:
  title: Noema API
  description: Production-grade AI reasoning engine
  version: "1.0.0"
  contact:
    name: Noema
servers:
  - url: http://localhost:8000
    description: Local development
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Master API key. Empty = auth disabled (dev only).
  schemas:
...
    InclusionProof:
      type: object
      required: [leaf_hash, path, root_hash, block_index]
      properties:
        leaf_hash:
          type: string
          description: SHA-256 hex digest of the serialized leaf data
        path:
          type: array
          description: Sibling hashes with direction (false = left, true = right)
          items:
            type: array
            minItems: 2
            maxItems: 2
            items:
              - type: string
              - type: boolean
        root_hash:
          type: string
          description: Merkle root this proof commits to
        block_index:
          type: integer
          description: Position of the leaf in the audit log
        leaf_data:
          type: object
          description: Original event payload (returned by the proof endpoint)
      example:
        leaf_hash: bf4f09dc175d3458bc8de84191cd18d95a5d21e97ebc8310f2b2d4a2c6d84b5e
        path: [["38a0d95e93f7bfc458f3b43bcf1034a7ad5c75adca7eff36c891ccc5f79c89e8", false]]
        root_hash: 28edc6ecf1c584c4ffd90d18565f7dc36a7e74c797cdbe57b4b1786a2f9c53f1
        block_index: 2
    Problem:
      type: object
      required: [type, title, status, detail]
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string

    # ── Request Bodies ──────────────────────────────────────────────
...
security:
  - ApiKeyAuth: []
paths:
...
  /admin/tasks/history:
    get:
      tags: [Admin]
      summary: Task history
      description: Return task history from the audit logger.
      operationId: adminTaskHistory
      parameters:
        - name: tenant_id
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 50
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: Task history events
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
        "503":
          description: Audit logger not available
...
  /admin/audit/proof/{task_id}:
    get:
      tags: [Admin]
      summary: Audit inclusion proof
      description: |
        Cryptographic Merkle inclusion proof for a completed task.
        The client can verify it against the published root hash
        without trusting the server or downloading the full audit log.
      operationId: adminAuditProof
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
          description: Task identifier
        - name: tenant_id
          in: query
          required: false
          schema:
            type: string
          description: Tenant identifier (falls back to current tenant)
      responses:
        "200":
          description: Inclusion proof
          content:
            application/json:
              schema:
                type: object
                properties:
                  tenant_id:
                    type: string
                  task_id:
                    type: string
                  proof:
                    $ref: "#/components/schemas/InclusionProof"
        "404":
          description: Task not found in audit log
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "503":
          description: Audit logger not available
  /admin/audit/verify:
    post:
      tags: [Admin]
      summary: Verify inclusion proof
      description: |
        Client-side verification of an inclusion proof. Returns whether the
        leaf data is cryptographically included in the Merkle tree for the
        given root hash. Does not trust the server.
      operationId: adminAuditVerify
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                proof:
                  $ref: "#/components/schemas/InclusionProof"
                leaf_data:
                  type: object
                  description: Original event payload (embedded in proof if omitted)
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
                  block_index:
                    type: integer
                  root_hash:
                    type: string
        "422":
          description: Malformed proof
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"

  # ── Features & Webhooks ────────────────────────────────────────
...